wip: #60 moved all tickets page
