+++ apps/issues/app/tickets/all/page.tsx
@@ -2,39 +2,44 @@ import { PlusIcon } from '@heroicons/react/outline';
 import type { Session } from 'next-auth';
 import Link from 'next/link';
 
-import ProjectList from '../../../components/ProjectList';
+import TicketList from '../../../components/TicketList';
 import { prisma } from '../../../lib/db';
 import { getSession } from '../../../lib/session';
 
-async function getProjects(session: Session) {
-  const projects = await prisma.project.findMany({
-    include: { team: { include: { user: true } }, tickets: true },
+async function getTickets(session: Session) {
+  const tickets = await prisma.ticket.findMany({
+    include: {
+      project: true,
+      ticketPriority: true,
+      ticketStatus: true,
+      ticketType: true,
+    },
     where: {
-      organizationId: { equals: session.user.settings.organization },
+      project: {
+        organizationId: { equals: session.user.settings.organization },
+      },
     },
   });
 
-  return projects;
+  return tickets;
 }
 
 export default async function Page() {
   const session = await getSession();
-  const projects = await getProjects(session);
+  const tickets = await getTickets(session);
 
   return (
     <div className="auto-rows-min gap-4 grid grid-cols-4">
       <div className="col-span-4 flex items-center justify-between px-2">
-        <span className="font-medium text-2xl text-slate-900">
-          All Projects
-        </span>
-        <Link href="/projects/add">
+        <span className="font-medium text-2xl text-slate-900">All Tickets</span>
+        <Link href="/tickets/add">
           <button className="bg-blue-100 border-2 border-blue-400 flex items-center px-3 py-1 rounded-md space-x-2 text-blue-900">
             <PlusIcon className="h-3 w-3" />
-            <span>Add Project</span>
+            <span>Add Ticket</span>
           </button>
         </Link>
       </div>
-      <ProjectList projects={projects} />
+      <TicketList tickets={tickets} />
     </div>
   );
 }
